UI Improvements
DrugCell now indicates when to prepare and give the drug. It also changes color depending on the time remaining. Grey > 2m, yellow <2m, red <15s.

Now the user is alerted with vibration/sound when a drug reminder state changes from prep to give and also when the timer is about to expire(<15s)

General UI improvements were added to give the user better clarity before the simulation test.